Allow for bond constraints
If thet standard deviation in the bond length is very small, then integrator error would lead to large energies. Degrees of freedom which have very small standard deviations are now treated as constrained. They don't contribute to the partition function or the logpdf. When generating random variates, the mean value is returned.